Just saving first big refactor. Need more work to be fully wiki 1.29
