+++ i18n/en.json
@@ -1,0 +1,11 @@
+{
+    "@metadata": {
+        "authors": [
+            "Arnoud Roques",
+            "Pieter J. Kersten",
+            "Wolfgang Fahl",
+            "Frederic Planchon"
+        ]
+    },
+    "plantuml-desc": "Renders a UML model from text using PlantUML."
+}
